+++ guppylang-internals/src/guppylang_internals/cfg/builder.py
@@ -357,13 +357,22 @@ class CFGBuilder(AstVisitor[BB | None]):
         for item in node.items:
             item.context_expr, bb = ExprBuilder.build(item.context_expr, self.cfg, bb)
             modifiers.push(self._handle_withitem(item))
-        accumulated_flags = self.cfg.unitary_flags | modifiers.flags()
+        accumulated_flags = self.cfg.unitary_flags.accumulate(modifiers.flags())
+        if UnitaryFlags.Dagger in accumulated_flags:
+            # `original_ast_body` is only needed for checking invalid constructs under
+            # dagger, so we only need to save it if dagger is present.
+            original_ast_body = copy.deepcopy(node.body)
+        else:
+            original_ast_body = None
+
         cfg = CFGBuilder().build(node.body, True, self.globals, accumulated_flags)
         new_node = ModifiedBlock(
-            cfg=cfg,
-            modifiers=modifiers,
+            cfg,
+            modifiers,
             # we save the first modifier node for a better error rendering
-            first_modifier_node=node.items[0].context_expr,
+            node.items[0].context_expr,
+            accumulated_flags,
+            original_ast_body,
             **dict(ast.iter_fields(node)),
         )