+++ guppylang-internals/src/guppylang_internals/definition/value.py
@@ -41,7 +41,7 @@ class CallableDef(ValueDef):
 
     @abstractmethod
     def check_call(
-        self, args: list[ast.expr], ty: Type, node: AstNode, ctx: "Context"
+        self, args: list[ast.expr], ty: Type, node: ast.Call, ctx: "Context"
     ) -> tuple[ast.expr, Subst]:
         """Checks the return type of a function call against a given type."""
 
@@ -57,8 +57,6 @@ class CallableDef(ValueDef):
 
 class CompiledCallableDef(CallableDef, CompiledValueDef):  # type: ignore[misc, unused-ignore]
     """Abstract base class a global module-level function."""
-
-    ty: FunctionType
 
     @abstractmethod
     def compile_call(