+++ tests/error/modifier_errors/dagger_loop2.py
@@ -1,0 +1,11 @@
+from guppylang.decorator import guppy
+from guppylang.std.builtins import dagger
+
+
+@guppy
+def test(i: int) -> None:
+    with dagger:
+        while i < 46:
+            pass
+
+test.compile()